+++ gocollect-client/shcollectors/find_test.go
@@ -12,7 +12,7 @@ func ExampleFind() {
 	// collectors.Runnable() has all keys of runnable collectors.
 	runnableCount := len(collectors.Runnable())
 	if runnableCount < 5 {
-		fmt.Print("strange, very few runners?")
+		fmt.Println("strange, very few runners?")
 	}
 
 	// We can run a single collector using those keys. For instance the
@@ -20,9 +20,9 @@ func ExampleFind() {
 	data := collectors.Run("core.id")
 	ip4 := data.GetString("ip4")
 	if ip4 == "" {
-		fmt.Print("ip4 empty?")
+		fmt.Println("ip4 empty?")
 	}
 
-	fmt.Print("runnables found")
+	fmt.Println("runnables found")
 	// Output: runnables found
 }